feat: added footer and input in replay mode
