+++ src/components/Chat/ChatInputFooter.tsx
@@ -1,0 +1,9 @@
+import { FooterMessage } from './FooterMessage';
+
+export const ChatInputFooter = () => {
+  return (
+    <div className="p-5 max-md:hidden">
+      <FooterMessage />
+    </div>
+  );
+};
